+++ backend/internal/dto/project.go
@@ -7,7 +7,7 @@ type (
 	Project struct {
 		ID          int       `json:"id"`
 		Name        string    `json:"name"`
-		LastRef     int16     `json:"last_ref"`
+		LastRef     int32     `json:"last_ref"`
 		Created     time.Time `json:"created"`
 		Modified    time.Time `json:"modified"`
 		ChangeCount int       `json:"change_count"`
